requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-ollama>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
datasets>=2.14.0
transformers>=4.30.0
accelerate>=0.20.0
torch>=2.0.0
tiktoken>=0.4.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
